Add PTZ Status Service.
This service allow to perform polling about PTZ status and to be notified
by providing our IPTZListener implementation. You can configure the initial
delay and the polling period. You have to manually start the service.
You can perform suspend and resume request.

Using a PTZ service you can implement, for example, a rendering of a AFOV with
azimuth and elevation.